holy shit that was painful
